Create a translation function to avoid the use of multiple files
The texts.js file was created to keep all the translations in the same place and pass the appropriate text to the Banner component according to the selected language. All the texts were saved in arrays and are selected by position. A better approach may be chosen in future updates.

The Banner component was changed to receive the language passed down by the App file as a prop and send it down the line to the selection text function, along with the text position indicating which text should be exhibit on screen.

The App.js file now only verifies the language in the translate button, instead of the whole app, since the use of files fully translated will be deprecated.

As a temporary step, the App.js is still showing the other components in the old translate files format.